requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.7
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
massedit==0.68.5
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
PyYAML==3.13
requests==2.20.1
six==1.12.0
tabulate==0.8.5
texttable==0.9.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
